pyproject.toml
[tool.poetry]
name = "homework14"
version = "0.1.0"
description = ""
package-mode = false
authors = ["Zdorovets Boris <bor.zdorovets28@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.0"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.34"
psycopg2 = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.9.1"}
requests = "^2.32.3"
libgravatar = "^1.0.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
python-dotenv = "^1.0.1"
alembic = "^1.13.2"
fastapi-mail = "^1.4.1"
redis = "^5.0.8"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.41.0"
pytest = "^8.3.3"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
